+++ server/src/api/controllers/RedirectShortUrl.controller.ts
@@ -5,6 +5,8 @@ import IRequest from "./interfaces/IRequest";
 import IReqParams from "./interfaces/IReqParams";
 import IMessage from "./interfaces/IMessage";
 import IResponse from "./interfaces/IResponse";
+import ErrorHandler from "../middleware/Handler/ErrorHandler";
+import HttpException from "../middleware/Exceptions/HttpException";
 
 const RedirectShortUrl = async (req: IRequest, res: Response) => {
   const msg: IMessage = { isSuccessful: false, message: "", info: null };
@@ -34,13 +36,8 @@ const RedirectShortUrl = async (req: IRequest, res: Response) => {
     }
   } catch (error) {
     if (error instanceof Error) {
-      msg.isSuccessful = false;
-      result.code = 500;
-      msg.message = error.name;
-      msg.info = error.message;
+      ErrorHandler(error as HttpException, res);
     }
-
-    return res.status(result.code).json(result.data);
   }
 };